+++ sep490-enterprise/src/main/java/greenbuildings/enterprise/services/WalletService.java
@@ -2,5 +2,5 @@ package greenbuildings.enterprise.services;
 
 
 public interface WalletService {
-    long getBalance();
+    double getBalance();
 }
